Support receiving duplicates of non-local items
The lines in ItemsReceived.txt are now comma separated and there are two
additional columns.

The first new column is the item index from Archipelago which allows for
games/clients to know the order that items have been received in or to
request a resync with the server if some items are never received.

Using the item index to tell when an item is new is required because the
received item IDs are no longer unique.

The second new column specifies whether the item has been received from
a non-local source (a different world or from a cheat command).

When a non-local item is received, and it belongs to a class of items
which allows for duplicates to be received, such as PsiCards, a new
unique copy is created and given to the player without writing to save
data that the item was received. Only the effect of receiving the item
is written to save data.

When a non-local item is received, but it belongs to a class of items
which does not allow for duplicates to be received, such as the Cobweb
Duster, the non-local item is treated as if it was a locally received
item instead. This means that if a Cobweb Duster is placed locally and a
non-local Cobweb Duster is received, then the local Cobweb Duster will
be collected too. Any extra Cobweb Duster items received after the first
will then be ignored because the first received Cobweb Duster will have
written to save data that it has been collected.

Gloria's Theater Candles are a special case because while multiple can
be received, there is a limit of 2. Receiving a non-local Candle will
treat the item as if it was locally received, but will prioritize
"Candle2" first because it is the least likely to have been placed
locally.

There are no changes to the behaviour of receiving local items.

The changes in this patch require similar changes to the AP client to
write received items in the new format.